+++ packages/studio-be/src/core/dialog/flow/flow-service.ts
@@ -17,7 +17,8 @@ import { Memoize } from 'lodash-decorators'
 import LRUCache from 'lru-cache'
 import moment from 'moment'
 import ms from 'ms'
-import nanoid from 'nanoid/generate'
+import { NLUService } from 'studio/nlu'
+import { QNAService } from 'studio/qna'
 
 import { validateFlowSchema } from '../utils/validator'
 
@@ -66,7 +67,9 @@ export class FlowService {
     @inject(TYPES.ObjectCache) private cache: ObjectCache,
     @inject(TYPES.KeyValueStore) private kvs: KeyValueStore,
     @inject(TYPES.BotService) private botService: BotService,
-    @inject(TYPES.JobService) private jobService: JobService
+    @inject(TYPES.JobService) private jobService: JobService,
+    @inject(TYPES.QnaService) private qnaService: QNAService,
+    @inject(TYPES.NLUService) private nluService: NLUService
   ) {
     this._listenForCacheInvalidation()
   }
@@ -108,6 +111,8 @@ export class FlowService {
         this.kvs.forBot(botId),
         this.logger,
         this.botService,
+        this.qnaService,
+        this.nluService,
         (key, flow, newKey) => this.invalidateFlow(botId, key, flow, newKey)
       )
       this.scopes[botId] = scope
@@ -126,6 +131,8 @@ export class ScopedFlowService {
     private kvs: KvsService,
     private logger: Logger,
     private botService: BotService,
+    private qnaService: QNAService,
+    private nluService: NLUService,
     private invalidateFlow: (key: string, flow?: FlowView, newKey?: string) => void
   ) {
     this.cache = new ArrayCache<string, FlowView>(
@@ -380,6 +387,12 @@ export class ScopedFlowService {
       this.ghost.renameFile(FLOW_DIR, previousUiName, newUiName)
     ])
 
+    await this.qnaService.onFlowRenamed({
+      botId: this.botId,
+      previousFlowName: previousName,
+      nextFlowName: newName
+    })
+
     await coreActions.onModuleEvent('onFlowRenamed', {
       botId: this.botId,
       previousFlowName: previousName,
@@ -445,6 +458,7 @@ export class ScopedFlowService {
     }
 
     if (!isNew) {
+      await this.qnaService.onFlowChanged({ botId: this.botId, flow })
       await coreActions.onModuleEvent('onFlowChanged', { botId: this.botId, flow })
     }
 
@@ -484,7 +498,13 @@ export class ScopedFlowService {
     topics = topics.filter(x => x.name !== topicName)
 
     await this.ghost.upsertFile('ndu', 'topics.json', JSON.stringify(topics, undefined, 2))
-    await coreActions.onModuleEvent('onTopicChanged', { botId: this.botId, oldName: topicName, newName: undefined })
+
+    const topicChanged = { botId: this.botId, oldName: topicName, newName: undefined }
+    await this.qnaService.onTopicChanged(topicChanged)
+    await this.nluService.onTopicChanged(topicChanged)
+
+    // TODO remove eventually
+    await coreActions.onModuleEvent('onTopicChanged', topicChanged)
   }
 
   public async createTopic(topic: Topic) {
@@ -492,7 +512,13 @@ export class ScopedFlowService {
     topics = _.uniqBy([...topics, topic], x => x.name)
 
     await this.ghost.upsertFile('ndu', 'topics.json', JSON.stringify(topics, undefined, 2))
-    await coreActions.onModuleEvent('onTopicChanged', { botId: this.botId, oldName: undefined, newName: topic.name })
+
+    const topicChanged = { botId: this.botId, oldName: undefined, newName: topic.name }
+    await this.qnaService.onTopicChanged(topicChanged)
+    await this.nluService.onTopicChanged(topicChanged)
+
+    // TODO remove eventually
+    await coreActions.onModuleEvent('onTopicChanged', topicChanged)
   }
 
   public async updateTopic(topic: Topic, topicName: string) {
@@ -502,7 +528,10 @@ export class ScopedFlowService {
     await this.ghost.upsertFile('ndu', 'topics.json', JSON.stringify(topics, undefined, 2))
 
     if (topicName !== topic.name) {
-      await coreActions.onModuleEvent('onTopicChanged', { botId: this.botId, oldName: topicName, newName: topic.name })
+      const topicChanged = { botId: this.botId, oldName: topicName, newName: topic.name }
+      await this.nluService.onTopicChanged(topicChanged)
+      await this.qnaService.onTopicChanged(topicChanged)
+      await coreActions.onModuleEvent('onTopicChanged', topicChanged)
 
       const flows = await this.loadAll()