+++ packages/ui-shared/src/Dialog/style.scss.d.ts
@@ -3,7 +3,6 @@
 interface CssExports {
   'dialog': string;
   'dialogBody': string;
-  'form': string;
 }
 declare var cssExports: CssExports;
 export = cssExports;